change type of n_sigmas in tof projectors to float

+++ include/tof_utils.h
@@ -15,7 +15,7 @@ void relevant_tof_bins(float x_m0,
 		       float tofbin_width,
 		       float tofcenter_offset,
 		       float sigma_tof,
-		       unsigned int n_sigmas,
+		       float n_sigmas,
 		       unsigned int n_half,
 		       int *it1,
 		       int *it2);
